Adjust our tests due to changes in JUnit 5 regarding inheritance
JUnit 5 changed how they handle inheritance to be more in line with what
standard Java does.
This change causes some issues with some of our tests as we can't
leverage overriding for static fields as we did before.

https://junit.org/junit5/docs/5.11.1/release-notes/#release-notes-5.11.0-junit-platform-bug-fixes
