+++ extensions/resteasy-reactive/rest-jackson/deployment/src/test/java/io/quarkus/resteasy/reactive/jackson/deployment/test/NonEmptyReflectionFreeSerializationTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 import io.quarkus.test.QuarkusUnitTest;
 
-public class NonEmptyReflectionFreeSerializationTest extends NonEmptySerializationTest {
+public class NonEmptyReflectionFreeSerializationTest extends AbstractNonEmptySerializationTest {
 
     @RegisterExtension
     static QuarkusUnitTest test = new QuarkusUnitTest()
